refactor(state): Index subs under 'display_name'
Index subreddits & subscriptions under display_name instead of url.

The display_name comes out to something like: 'news', whereas the url
is: '/r/news'. We already know that it's a subreddit with a url shaped
like that, due to the context. Also, it's more consistent when seeing a
post, which references the subreddit by its display_name.

fix(state): Fix mapping of popular subreddits

Map popular subreddits by display_name. Was coming back as an array.