feat(state): Expand subreddits, see below
Within the subreddits branch of state:

- 'collectionToShow' can now swap out popular subreddits for
  subscriptions and vice versa.
- 'allSubs': stores all the subs, whether or not user is subscribed.

Was having a problem with interacting with 'subscriptions.subscribedTo',
where the first key is '[FRONT_PAGE]'. So just removed it. It won't
show up in the Dropdown list anymore (Feed/SelectSubreddit).

Helper method for receiving subreddits was refactored, because
subscriptions & popular each need separate keys, when piping info to
the showSubredditCollection action creator.