refactor(state/actions/feed): Modify src code for testability
Need to return a rejected promise in order to catch it in the test

refactor(state/actions/feed): Split up 'toggleMultipleMode' a.c. into two

The old way mutated state mid-way through, which made it impossible to
test with the current test tool setup. It's much easier to refactor the
action creator to make it more testable, than it is to change any
mocking/stubbing configuration in tests.

It is now up to the component code to compose a chain of action
creators, because 'toggleMultipleMode' does much less than before.